refactor: move AutoConfiguration package to autoconfigure.application namespace
- io.github.ppzxc.template.application.autoconfiguration
  → io.github.ppzxc.template.autoconfigure.application
- Follows ADR-0014: AutoConfiguration classes use autoconfigure.* namespace
