+++ template/template-application-autoconfiguration/src/main/java/io/github/ppzxc/template/autoconfigure/application/ApplicationAutoConfiguration.java
@@ -1,4 +1,4 @@
-package io.github.ppzxc.template.application.autoconfiguration;
+package io.github.ppzxc.template.autoconfigure.application;
 
 import org.springframework.boot.autoconfigure.AutoConfiguration;
