+++ template/template-application-autoconfiguration/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.ppzxc.template.application.autoconfiguration.ApplicationAutoConfiguration
+io.github.ppzxc.template.autoconfigure.application.ApplicationAutoConfiguration
